Separate test logic for sql packagesf
