+++ tracer/contrib/sqltraced/pq_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/DataDog/dd-trace-go/tracer"
-	"github.com/DataDog/dd-trace-go/tracer/contrib/sqltraced/sqlutils"
+	"github.com/DataDog/dd-trace-go/tracer/contrib/sqltraced/sqltest"
 	"github.com/DataDog/dd-trace-go/tracer/tracertest"
 	"github.com/lib/pq"
 )
@@ -18,7 +18,7 @@ func TestPostgres(t *testing.T) {
 	}
 	defer db.Close()
 
-	testDB := &sqlutils.DB{
+	testDB := &sqltest.DB{
 		DB:         db,
 		Tracer:     trc,
 		Transport:  transport,
@@ -37,5 +37,5 @@ func TestPostgres(t *testing.T) {
 		"db.name":  "circle_test",
 	}
 
-	sqlutils.AllSQLTests(t, testDB, expectedSpan)
+	sqltest.AllSQLTests(t, testDB, expectedSpan)
 }
