+++ tracer/contrib/sqltraced/sqltest/world.mysql.sql
