+++ tracer/contrib/sqltraced/sqlutils/sqlutils.go
@@ -1,2 +1,2 @@
-// Package sqlutils share some utils functions for sql testing
+// Package sqlutils share some utils functions for sql packages
 package sqlutils
