+++ tracer/contrib/sqlxtraced/mysql_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/DataDog/dd-trace-go/tracer"
-	"github.com/DataDog/dd-trace-go/tracer/contrib/sqltraced/sqlutils"
+	"github.com/DataDog/dd-trace-go/tracer/contrib/sqltraced/sqltest"
 	"github.com/DataDog/dd-trace-go/tracer/tracertest"
 	"github.com/go-sql-driver/mysql"
 )
@@ -18,7 +18,7 @@ func TestMySQL(t *testing.T) {
 	}
 	defer dbx.Close()
 
-	testDB := &sqlutils.DB{
+	testDB := &sqltest.DB{
 		DB:         dbx.DB,
 		Tracer:     trc,
 		Transport:  transport,
@@ -37,5 +37,5 @@ func TestMySQL(t *testing.T) {
 		"db.name":  "circle_test",
 	}
 
-	sqlutils.AllSQLTests(t, testDB, expectedSpan)
+	sqltest.AllSQLTests(t, testDB, expectedSpan)
 }
